bugfix/ETD-853-connectors-the-binance-api-will-

+++ plugin-base/src/main/java/io/everytrade/server/plugin/impl/everytrade/BinanceDownloader.java
@@ -94,7 +94,7 @@ public class BinanceDownloader {
                     break;
                 }
                 userTrades.addAll(userTradesBlock);
-                lastDownloadedTx = userTradesBlock.stream().max(comparing(u -> Integer.parseInt(u.getId()))).get().getId();
+                lastDownloadedTx = userTradesBlock.stream().max(comparing(u -> Long.parseLong(u.getId()))).get().getId();
                 params.setStartId(lastDownloadedTx);
             }
             currencyPairLastIds.put(pair.toString(), lastDownloadedTx);
